project/requirements_without_dep.txt
Django==4.1
psycopg2-binary==2.9.3
freezegun==1.2.2
django-jalali==6.0.0
django-admin-sortable2==2.1.2
django-environ==0.9.0
persiantools==3.0.1
django-ckeditor==6.5.1

#ipython==8.4.0
#django-extensions==3.2.0
